Create C "Hello, World!" Program.c
The #include is a preprocessor command that tells the compiler to include the contents of stdio.h (standard input and output) file in the program.
The stdio.h file contains functions such as scanf() and printf() to take input and display output respectively.
If you use the printf() function without writing #include <stdio.h>, the program will not compile.
The execution of a C program starts from the main() function.
printf() is a library function to send formatted output to the screen. In this program, printf() displays Hello, World! text on the screen.
The return 0; statement is the "Exit status" of the program. In simple terms, the program ends with this statement.