Create C Program to Compute Quotient and Remainder.c
In this program, the user is asked to enter two integers (dividend and divisor). They are stored in variables dividend and divisor respectively.
Then the quotient is evaluated using / (the division operator), and stored in quotient.
Similarly, the remainder is evaluated using % (the modulo operator) and stored in remainder.
Finally, the quotient and remainder are displayed using printf().